fix Lua3rdMeta
把Lua3rdMeta独立成一个文件，否则打包到手机上读取scriptableobject的时候会出错
