+++ Assets/Plugins/Slua_Managed/Lua3rdMeta.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: b51a3b5f40c08405ca2dfb572d3558c0
+timeCreated: 1451443061
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
